Moved staxUser related classes to database module
