+++ app/src/main/java/com/hover/stax/contacts/ContactDao.kt
@@ -19,7 +19,7 @@ import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Update
-import com.hover.stax.data.local.BaseDao
+import com.hover.stax.database.dao.BaseDao
 
 @Dao
 interface ContactDao : BaseDao<StaxContact> {
