+++ app/src/main/java/com/hover/stax/database/AppDatabase.kt
@@ -25,9 +25,9 @@ import com.hover.stax.contacts.ContactDao
 import com.hover.stax.contacts.StaxContact
 import com.hover.stax.data.local.accounts.AccountDao
 import com.hover.stax.data.local.channels.ChannelDao
-import com.hover.stax.data.local.user.UserDao
+import com.hover.stax.database.dao.user.UserDao
 import com.hover.stax.domain.model.Account
-import com.hover.stax.domain.model.StaxUser
+import com.hover.stax.database.entity.StaxUser
 import com.hover.stax.merchants.Merchant
 import com.hover.stax.merchants.MerchantDao
 import com.hover.stax.paybill.Paybill
