+++ app/src/main/java/com/hover/stax/domain/use_case/stax_user/StaxUserUseCase.kt
@@ -15,12 +15,12 @@
  */
 package com.hover.stax.domain.use_case.stax_user
 
-import com.hover.stax.domain.model.StaxUser
-import com.hover.stax.domain.repository.StaxUserRepository
+import com.hover.stax.database.entity.StaxUser
+import com.hover.stax.database.repository.StaxUserRepository
 
 class StaxUserUseCase(private val staxUserRepository: StaxUserRepository) {
 
-    val user = staxUserRepository.staxUser
+    val user = staxUserRepository.getUserAsync()
 
     suspend fun saveUser(user: StaxUser) = staxUserRepository.saveUser(user)
