+++ app/src/main/java/com/hover/stax/paybill/PaybillDao.kt
@@ -19,7 +19,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.hover.stax.data.local.BaseDao
+import com.hover.stax.database.dao.BaseDao
 import kotlinx.coroutines.flow.Flow
 
 @Dao
