+++ app/src/main/java/com/hover/stax/requests/RequestDao.kt
@@ -21,7 +21,7 @@ import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
-import com.hover.stax.data.local.BaseDao
+import com.hover.stax.database.dao.BaseDao
 
 @Dao
 interface RequestDao : BaseDao<Request> {
