+++ app/src/main/java/com/hover/stax/schedules/ScheduleDao.kt
@@ -20,7 +20,7 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Query
 import androidx.room.Update
-import com.hover.stax.data.local.BaseDao
+import com.hover.stax.database.dao.BaseDao
 
 @Dao
 interface ScheduleDao : BaseDao<Schedule> {
