+++ app/src/main/java/com/hover/stax/transactions/TransactionDao.kt
@@ -21,7 +21,7 @@ import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
-import com.hover.stax.data.local.BaseDao
+import com.hover.stax.database.dao.BaseDao
 import kotlinx.coroutines.flow.Flow
 
 @Dao
